+++ README.md
@@ -5,16 +5,26 @@
 前回のカレンダーver.3に予定の編集更新、削除機能を追加したもの。
 
 ## 工夫した点・こだわった点
-- 登録された予定を呼び出し、編集画面を表示して編集更新する。単純な予定の更新以外に、繰り返し予定として登録されたものは、編集画面にその日だけ変更するか、その日以降の予定をすべて変更するか、の選択肢を表示させ、条件分岐してデータを更新するようにした。
+- 登録された予定を呼び出し、編集画面を表示して編集更新する。単純な予定の更新以外に、
+繰り返し予定として登録されたものは、編集画面にその日だけ変更するか、その日以降の予定をすべて変更するか、
+の選択肢を表示させ、条件分岐してデータを更新するようにした。
 - 削除機能も編集更新と同様に、繰り返し予定の場合は、その日のみ削除か、その日以降すべて削除か、条件分岐できるようにした。
-- 前回追加した機能「予定検索」で検索した結果に対しても、結果の予定をクリックすると、その画面から編集画面に飛んで編集更新、削除ができるようにした。
+- 前回追加した機能「予定検索」で検索した結果に対しても、結果の予定をクリックすると、
+その画面から編集画面に飛んで編集更新、削除ができるようにした。
 - 結果的に、すべてひとつのモーダル画面を共用して、ボタン選択だけで予定の登録、更新、削除ができるようになった。
+- さすがにmain.jsのコードが長くなってきたので、いくつかのファイルに分轄した。
+
 
 ## 次回トライしたこと（または機能）
 - ユーザー認証機能
 - 画像の登録や日記への発展なども検討したい
 
 ## 備考（感想、シェアしたいこと等なんでも）
-- すでに登録している予定の編集更新、削除は容易そうに思えるかもしれないが、結構難しい。繰り返しの予定の編集には大変苦労した。
-- main.jsから$.post()メソッドでupdate.phpにデータを送る仕様になっているが、update.phpでvar_dumpすると、$_POSTの中身が常に空になっている現象に悩まされ、ついにXAMPPを再インストールするに至るも解決せず、ほぼ1日の時間を大きく浪費してしまったのが誠に悔しい。
+- すでに登録している予定の編集更新、削除は容易そうに思えるかもしれないが、結構難しい。
+繰り返しの予定の編集には大変苦労した。
+- main.jsから$.post()メソッドでupdate.phpにデータを送る仕様になっているが、
+update.phpでvar_dumpすると、$_POSTの中身が常に空になっている現象に悩まされ、
+ついにXAMPPを再インストールするに至るも解決せず、ほぼ1日の時間を大きく浪費してしまったのが誠に悔しい。
+- jsのコードが長くなって複雑にからみあっているので、それを解いて分割する作業は大変苦労した。
+新規作成時からある程度のところで順次、分けていくとスムーズかと思える。
